Move sources to PointwiseFunctions
