+++ src/Evolution/Executables/ScalarTensor/ScalarTensorBase.hpp
@@ -47,8 +47,6 @@
 #include "Evolution/Systems/ScalarTensor/BoundaryCorrections/Factory.hpp"
 #include "Evolution/Systems/ScalarTensor/Constraints.hpp"
 #include "Evolution/Systems/ScalarTensor/Initialize.hpp"
-#include "Evolution/Systems/ScalarTensor/Sources/ScalarSource.hpp"
-#include "Evolution/Systems/ScalarTensor/StressEnergy.hpp"
 #include "Evolution/Systems/ScalarTensor/System.hpp"
 #include "Evolution/Systems/ScalarTensor/Tags.hpp"
 #include "Evolution/Tags/Filter.hpp"
@@ -129,6 +127,8 @@
 #include "PointwiseFunctions/ScalarTensor/ConstraintGammas.hpp"
 #include "PointwiseFunctions/ScalarTensor/ScalarCharge.hpp"
 #include "PointwiseFunctions/ScalarTensor/SourceTags.hpp"
+#include "PointwiseFunctions/ScalarTensor/ScalarSource.hpp"
+#include "PointwiseFunctions/ScalarTensor/StressEnergy.hpp"
 #include "Time/Actions/AdvanceTime.hpp"
 #include "Time/Actions/CleanHistory.hpp"
 #include "Time/Actions/RecordTimeStepperData.hpp"
