+++ src/Evolution/Systems/ScalarTensor/CMakeLists.txt
@@ -8,7 +8,6 @@ add_spectre_library(${LIBRARY})
 spectre_target_sources(
   ${LIBRARY}
   PRIVATE
-  StressEnergy.cpp
   TimeDerivative.cpp
   VolumeTermsInstantiation.cpp
   )
@@ -20,7 +19,6 @@ spectre_target_headers(
   Characteristics.hpp
   Constraints.hpp
   Initialize.hpp
-  StressEnergy.hpp
   System.hpp
   Tags.hpp
   TimeDerivative.hpp
@@ -37,6 +35,7 @@ target_link_libraries(
   GeneralRelativity
   GeneralizedHarmonic
   GhScalarTensorAnalyticData
+  ScalarTensorPointwise
   Serialization
   Utilities
   INTERFACE
@@ -47,4 +46,3 @@ target_link_libraries(
 add_subdirectory(Actions)
 add_subdirectory(BoundaryConditions)
 add_subdirectory(BoundaryCorrections)
-add_subdirectory(Sources)
