--- src/Evolution/Systems/ScalarTensor/Sources/CMakeLists.txt
@@ -1,15 +1,0 @@
-# Distributed under the MIT License.
-# See LICENSE.txt for details.
-
-spectre_target_sources(
-  ${LIBRARY}
-  PRIVATE
-  ScalarSource.cpp
-)
-
-spectre_target_headers(
-  ${LIBRARY}
-  INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
-  HEADERS
-  ScalarSource.hpp
-)
